feat(modules): Bans (yolo)

+++ Admins.Core/src/Admins/API.cs
@@ -13,16 +13,20 @@ namespace Admins.Core.Admins;
 public class AdminsManager : IAdminsManager
 {
     private ISwiftlyCore _core;
-    private ServerAdmins _serverAdmins;
+    private ServerAdmins? _serverAdmins;
     private IOptionsMonitor<CoreConfiguration>? _confMonitor;
 
     public event Action<IPlayer, IAdmin>? OnAdminLoad;
 
-    public AdminsManager(ISwiftlyCore core, ServerAdmins serverAdmins, IOptionsMonitor<CoreConfiguration> confMonitor)
+    public AdminsManager(ISwiftlyCore core, IOptionsMonitor<CoreConfiguration> confMonitor)
     {
         _core = core;
-        _serverAdmins = serverAdmins;
         _confMonitor = confMonitor;
+    }
+
+    public void SetServerAdmins(ServerAdmins serverAdmins)
+    {
+        _serverAdmins = serverAdmins;
     }
 
     public IAdmin? AddAdmin(ulong steamId64, string adminName, List<IGroup> groups, List<string> permissions)
@@ -42,7 +46,7 @@ public class AdminsManager : IAdminsManager
             {
                 var db = _core.Database.GetConnection("admins");
                 await db.InsertAsync(newAdmin);
-                _serverAdmins.Load();
+                _serverAdmins?.Load();
             }
         });
 
@@ -80,19 +84,21 @@ public class AdminsManager : IAdminsManager
     {
         return ServerAdmins.AllAdmins.Values.Cast<IAdmin>().ToList();
     }
-
     public void RefreshAdmins()
     {
-        _serverAdmins.Load();
+        _serverAdmins?.Load();
     }
 
     public void RemoveAdmin(IAdmin admin)
     {
         Task.Run(async () =>
         {
-            var db = _core.Database.GetConnection("admins");
-            await db.DeleteAsync((Admin)admin);
-            _serverAdmins.Load();
+            if (_confMonitor!.CurrentValue.UseDatabase == true)
+            {
+                var db = _core.Database.GetConnection("admins");
+                await db.DeleteAsync((Admin)admin);
+            }
+            _serverAdmins?.Load();
         });
     }
 
@@ -105,14 +111,13 @@ public class AdminsManager : IAdminsManager
     {
         ServerAdmins.AllAdmins = new ConcurrentDictionary<ulong, Admin>(admins.Cast<Admin>().ToDictionary(a => (ulong)a.SteamId64, a => a));
     }
-
     public void AssignAdmin(IPlayer player, IAdmin admin)
     {
-        _serverAdmins.AssignAdmin(player, (Admin)admin);
+        _serverAdmins?.AssignAdmin(player, (Admin)admin);
     }
 
     public void UnassignAdmin(IPlayer player, IAdmin admin)
     {
-        _serverAdmins.UnassignAdmin(player, (Admin)admin);
+        _serverAdmins?.UnassignAdmin(player, (Admin)admin);
     }
 }
